all ppi retreivals now working
