+++ src/pyhail/mesh_ppi.py
@@ -6,22 +6,93 @@ Joshua Soderholm - 15 August 2020
 """
 
 import numpy as np
-from pyhail import common
+print('imports static in mesh')
+import common
+
+def _antenna_to_arc(ranges, elevation):
+    """
+    Return the great circle distance directly below the radar beam and the
+    altitude of the radar beam.
+    ----------
+    ranges : 1d array
+        Distances to the center of the radar gates (bins) in meters.
+    elevation : float
+        Elevation angle of the radar in degrees.
+    Returns
+    -------
+    s: 1d array
+        Distance along the great circle for each radar bin (units: meters)
+    z: 1d array
+        Altitude above radar level for each radar bin (units: meters)
+    Notes
+    -----
+    The calculation for Cartesian coordinate is adapted from equations
+    2.28(b) and 2.28(c) of Doviak and Zrnic [1]_ assuming a
+    standard atmosphere (4/3 Earth's radius model).
+    .. math::
+        z = \\sqrt{r^2+R^2+2*r*R*sin(\\theta_e)} - R
+        s = R * arcsin(\\frac{r*cos(\\theta_e)}{R+z})
+    Where r is the distance from the radar to the center of the gate,
+    :math:`\\theta_a` is the azimuth angle, :math:`\\theta_e` is the
+    elevation angle, s is the arc length, and R is the effective radius
+    of the earth, taken to be 4/3 the mean radius of earth (6371 km).
+    References
+    ----------
+    .. [1] Doviak and Zrnic, Doppler Radar and Weather Observations, Second
+        Edition, 1993, p. 21.
+    """
+    theta_e = elevation * np.pi / 180.0  # elevation angle in radians.
+    R = 6371.0 * 1000.0 * 4.0 / 3.0  # effective radius of earth in meters.
+    r = ranges
+    
+    z = (r ** 2 + R ** 2 + 2.0 * r * R * np.sin(theta_e)) ** 0.5 - R
+    s = R * np.arcsin(r * np.cos(theta_e) / (R + z))  # arc length in m.
+    return s, z
+
+def _calc_dz(column_z):
+    """
+    Calculate altitude difference between elements in a 1d array
+    Takes into account the boundaries
+
+    Parameters
+    ----------
+    column_z : 1darray
+        altitude of column samples
+    
+    Returns
+    -------
+    dz : 1darray
+        Difference between altitude elements
+    """
+
+    dz = []
+    n_ppi = len(column_z)
+    for i in range(n_ppi):
+        #calculate dz for use in shi calc
+        if i == 0:
+            value = column_z[i + 1] - column_z[i]
+        elif (i != 0) & (i != n_ppi - 1):
+            value = (column_z[i + 1] - column_z[i - 1]) / 2
+        else:
+            value = column_z[i] - column_z[i - 1]
+        dz.append(value)
+    dz = np.array(dz)
+    return dz
 
 def main(
-    data_dict,
-    coords_dict,
-    dbz_fname,
+    reflectivity,
+    elevation,
+    azimuth,
+    rangebin,
+    radar_altitude,
     levels,
-    radar_band='C',
+    radar_band='S',
     min_range=10,
     max_range=150,
-    mesh_method="mh2019_95",
-    mesh_fname=None,
-    posh_fname=None,
-    ke_fname=None,
-    shi_fname=None,
-    correct_cband_refl=True
+    mesh_method="mh2019_75",
+    correct_cband_refl=True,
+    minimum_sweeps_raise_expection=4,
+    minimum_sweeps_raise_warning=8
 ):
 
     """
@@ -29,12 +100,16 @@ def main(
 
     Parameters
     ----------
-    data_dict : dictionary
-        Dictionary containing radar fields, each field contains a list of sweeps, where each sweep is an 2D array.
-    coords_dict: dictionary
-        Containing coordinate fields
-    dbz_fname : str
-        Name of reflectivity field in the radar data_dict.
+    reflectivity : list of 2D ndarrays
+        list where each element is the sweep reflectivity data in an array with dimensions (azimuth, range)
+    elevation: 1d ndarray of floats
+        ndarray where each element is the fixed elevation angle of the sweep
+    azimuth: list of 1D ndarrays
+        list where each element is the sweep azimuth angles
+    rangebin: list of 1D ndarrays
+        list where each element is the sweep range distances
+    radar_altitude: float
+        altitude of radar AMSL
     levels : list of length 2
         height above sea level (m) of the freezing level and -20C level (in any order)
     radar_band: str 
@@ -43,13 +118,15 @@ def main(
         minimum surface range for MESH retrieval (m)
     max_range: int
         maximum surface range for MESH retrieval (m)
-    mesh_fname, posh_fname, ke_fname, shi_fname : str
-        String to name new hail field that will be added to the grid object.
-        Default is 'mesh', 'posh', 'hail_ke', 'shi'.
     mesh_method : string
         either witt1998, mh2019_75 or mh2019_95. see more information below
     correct_cband_refl: logical
         flag to trigger C band hail reflectivity correction (if radar_band is C)
+    minimum_sweeps_raise_expection: int
+        minimum number of sweeps to raise an exception
+    minimum_sweeps_raise_warning: int
+        minimum number of sweeps to raise a warning
+    
     Returns
     -------
     output_fields : dictionary
@@ -67,223 +144,175 @@ def main(
     Zl = 40
     Zu = 50
 
-    # default field names
-    if mesh_fname is None:
-        mesh_fname = "mesh_" + mesh_method
-    if posh_fname is None:
-        posh_fname = "posh"
-    if ke_fname is None:
-        ke_fname = "hail_ke"
-    if shi_fname is None:
-        shi_fname = "shi"
-
     # This dummy proofs the user input. The melting level will always
     # be lower in elevation than the negative 20 deg C isotherm
     meltlayer = np.min(levels)
     neg20layer = np.max(levels)
 
-    # Initialize dimensions
-    el = coords_dict['fixed_angle']
-    sort_idx = list(np.argsort(el))
-    el = el[sort_idx]
-    n_ppi = len(el)
-    az = common.get_field(0, coords_dict['azimuth'], coords_dict)
-    n_rays = len(az)
-    rg = coords_dict['range']
-    n_bins = len(rg)
+    # sort by fixed angle
+    sort_idx = list(np.argsort(elevation))
+    reflectivity_dataset = [reflectivity[i] for i in sort_idx]
+    elevation_dataset = [elevation[i] for i in sort_idx]
+    azimuth_dataset = [azimuth[i] for i in sort_idx]
+    range_dataset = [rangebin[i] for i in sort_idx] 
 
     # require more than one sweep
-    if len(el) <= 1:
-        raise Exception("Require more than one sweep to calculate MESH")
-    elif len(el) < 10:
-        raise Warning("Number of sweep is less than 10 and not recommended for MESH calculations")
+    if len(elevation_dataset) <= minimum_sweeps_raise_expection:
+        raise Exception(f"Require more than {minimum_sweeps_raise_expection} sweeps to calculate MESH")
+    elif len(elevation_dataset) < minimum_sweeps_raise_warning:
+        raise Warning("Number of sweep is less than {minimum_sweeps_raise_warning} and not recommended for MESH calculations")
     
-    # Initialize arrays
-    DBZ = np.zeros((len(el), len(az), len(rg)))
-    X = np.zeros_like(DBZ)
-    Y = np.zeros_like(DBZ)
-    Z = np.zeros_like(DBZ)
-    dZ = np.zeros_like(DBZ)
-    SHI = np.zeros((len(az), len(rg)))
-
-    #build coords
-    rg_grid, az_grid = np.meshgrid(rg, coords_dict['azimuth'])
-    rg_grid, elv_grid = np.meshgrid(rg, coords_dict['elevation'])
-    x_ppi, y_ppi, z_ppi = common.antenna_to_cartesian(rg_grid / 1000, az_grid, elv_grid)
-
-    # build 3D vol grids of reflectivity and Cartesian coords
-    for i, el_idx in enumerate(sort_idx):
-        tmp_field = common.get_field(el_idx, data_dict[dbz_fname], coords_dict)
-        if np.shape(tmp_field) == (len(az), len(rg)):
-            DBZ[i, :, :] = tmp_field
-        else:
-            raise Exception("Corrupt volume, sweeps of different shapes detected")
-        s = common.get_slice(el_idx, coords_dict)
-        X[i, :, :] = x_ppi[s]
-        Y[i, :, :] = y_ppi[s]
-        Z[i, :, :] = z_ppi[s] + coords_dict['altitude'] #units m at ASL required for NWP data
-    # calculate ground range by ignoring Z
-    ground_range = np.sqrt(X ** 2 + Y ** 2)
-
-    #apply C band correction
+    # Initialize sweep coords
+    sweep0_nrays = len(azimuth_dataset[0])
+    sweep0_nbins = len(range_dataset[0])
+    n_ppi = len(elevation_dataset)
+    z_dataset = [] #list (dim: elevation) of 1d array (dim: range) for each sweep, altitude above ground level (m) of each range bin
+    s_dataset = [] #list (dim: elevation) of 1d array (dim: range) for each sweep, great circle arc distance (m) of each radar bin
+    wt_dataset = [] #list (dim: elevation) of 1d array (dim: range) for each sweep, temperature weighting value
+    hail_ke_dataset = [] #list (dim: elevation) of 2d array (dim: azimuth, range) for each sweep, hail kinetic energy
     hail_refl_correction_description = ''
-    if radar_band == 'C' and correct_cband_refl:
-        DBZ = DBZ*1.113 - 3.929
-        hail_refl_correction_description = "C band hail reflectivity correction applied from Brook et al. 2023 https://arxiv.org/abs/2306.12016"
-
-    # calculate dZ (used for SHI)
     for i in range(n_ppi):
-        if i == 0:
-            dZ[i, :, :] = Z[i + 1, :, :] - Z[i, :, :]
-        if (i != 0) & (i != n_ppi - 1):
-            dZ[i, :, :] = (Z[i + 1, :, :] - Z[i - 1, :, :]) / 2
-        if i == n_ppi - 1:
-            dZ[i, :, :] = Z[i, :, :] - Z[i - 1, :, :]
+        #calculate cartesian coordinates
+        s, z = _antenna_to_arc(range_dataset[i], elevation_dataset[i])
+        s_dataset.append(s)
+        z_dataset.append(z + radar_altitude)
+        # calc temperature based weighting function
+        wt = (z_dataset[i] - meltlayer) / (neg20layer - meltlayer)
+        wt[z_dataset[i] <= meltlayer] = 0
+        wt[z_dataset[i] >= neg20layer] = 1
+        wt[wt < 0] = 0
+        wt[wt > 1] = 1
+        wt_dataset.append(wt)
+        #apply C band correction
+        if radar_band == 'C' and correct_cband_refl:
+            reflectivity_dataset[i] = reflectivity_dataset[i] * 1.113 - 3.929
+            hail_refl_correction_description = "C band hail reflectivity correction applied from Brook et al. 2023 https://arxiv.org/abs/2306.12016"
+        # calc weights for hail kenetic energy
+        reflectivity_weights = (reflectivity_dataset[i] - Zl) / (Zu - Zl)
+        reflectivity_weights[reflectivity_dataset[i] <= Zl] = 0
+        reflectivity_weights[reflectivity_dataset[i] >= Zu] = 1
+        reflectivity_weights[reflectivity_weights < 0] = 0
+        reflectivity_weights[reflectivity_weights > 1] = 1
+        #limit on DBZ
+        reflectivity_dataset[i][reflectivity_dataset[i] > 100] = 100
+        reflectivity_dataset[i][reflectivity_dataset[i] < -100] = -100
+        # calc hail kenetic energy
+        hail_ke = (5.0e-6) * 10 ** (0.084 * reflectivity_dataset[i]) * reflectivity_weights
+        hail_ke[np.isnan(hail_ke)] = 0
+        hail_ke_dataset.append(hail_ke)
 
-    # calc hail kenetic energy
-    DBZ_weights = (DBZ - Zl) / (Zu - Zl)
-    DBZ_weights[DBZ <= Zl] = 0
-    DBZ_weights[DBZ >= Zu] = 1
-    DBZ_weights[DBZ_weights < 0] = 0
-    DBZ_weights[DBZ_weights > 1] = 1
-    #limit on DBZ
-    DBZ[DBZ > 100] = 100
-    DBZ[DBZ < -100] = -100
+    #generate arc range and dz lookup (note these have different dimensions to the dimension variables)
+    dz_dataset = [] #list (dim: range) where each element represents a range bin, 1d array (dim: elevation) where each element represents a sweep, altitude dz for shi integration (m)
+    s_lookup_dataset = [] #list (dim: range) where each element represents a range bin, a 1d array (dim: elevation[1:]) of the index to use from each sweep above sweep0
+    for rg_idx in range(sweep0_nbins):
+        s_lookup = [0]
+        column_z = [z_dataset[0][rg_idx]]
+        for sweep_idx in range(1, n_ppi, 1):
+            closest_rng_idx = np.argmin(np.abs(s_dataset[0][rg_idx]-s_dataset[sweep_idx]))
+            s_lookup.append(closest_rng_idx)
+            column_z.append(z_dataset[sweep_idx][closest_rng_idx])
+        s_lookup_dataset.append(np.array(s_lookup))
+        dz_dataset.append(_calc_dz(column_z))
 
-    E = (5.0e-6) * 10 ** (0.084 * DBZ) * DBZ_weights
-
-    # calc temperature based weighting function
-    Wt = (Z - meltlayer) / (neg20layer - meltlayer)
-    Wt[Z <= meltlayer] = 0
-    Wt[Z >= neg20layer] = 1
-    Wt[Wt < 0] = 0
-    Wt[Wt > 1] = 1
-
-    # calc severe hail index (element wise for integration)
-    SHI_elements = Wt * E * dZ
-    # calc valid mask
-    valid = (
-        (Wt > 0)
-        & (E > 0)
-        & (ground_range > min_range * 1000)
-        & (ground_range < max_range * 1000)
-    )
-
-    # loop through each azimuth
-    for az_idx in range(n_rays):
-        slice_valid = valid[:, az_idx, :]
-        if not np.any(slice_valid):
-            continue
-        slice_SHI_elements = SHI_elements[:, az_idx, :]
-        # if there's samples that are valid, loop through each surface PPI range bin
-        for rg_idx in range(n_bins):
-            SHI_temp = 0
-            surface_rg_value = rg[rg_idx]
-            for el_idx in range(n_ppi):
-                # skip invalid
-                if not slice_valid[el_idx, rg_idx]:
-                    continue
-                # skip empty values
-                if slice_SHI_elements[el_idx, rg_idx] == 0:
-                    continue
-                # if lowest PPI (always index 0 in the 3D grid), just use the SHI value directory
-                if el_idx == 0:
-                    SHI_temp = slice_SHI_elements[el_idx, rg_idx]
-                else:
-                    # find the nearest element in range to the surface_rg_value in the current ray
-                    ppi_ray_rg = ground_range[el_idx, az_idx, :]
-                    closest_idx = np.argmin(np.abs(ppi_ray_rg - surface_rg_value))
-                    SHI_temp += slice_SHI_elements[el_idx, closest_idx]
+    #calculate shi on lowest sweep coordinates
+    shi = np.zeros((len(azimuth_dataset[0]), len(range_dataset[0])))
+    shi_mask = np.zeros((len(azimuth_dataset[0]), len(range_dataset[0])), dtype=bool)
+    # loop through each ray in the lowest sweep
+    for az_idx in range(sweep0_nrays):
+        #print(az_idx)
+        # loop through each range bin for the ray
+        for rg_idx in range(sweep0_nbins):
+            #check if sweep0 (lowest) range is outside of limits
+            if s_dataset[0][rg_idx] < min_range * 1000 or s_dataset[0][rg_idx] > max_range * 1000:
+                shi_mask[az_idx, rg_idx] = True
+                continue
+            #init shi elements
+            column_shi_elements = [hail_ke_dataset[0][az_idx, rg_idx] * wt_dataset[0][rg_idx]] # HKE * WT
+            #loop through other sweeps above ground
+            for sweep_idx in range(1, n_ppi, 1):
+                # find closest point using great circle arc to sweep0 (lowest) location
+                closest_idx = s_lookup_dataset[rg_idx][sweep_idx]
+                #append
+                column_shi_elements.append(hail_ke_dataset[sweep_idx][az_idx, closest_idx] * wt_dataset[sweep_idx][closest_idx])
             # insert into SHI if there's a valid value
-            if SHI_temp > 0:
-                SHI[az_idx, rg_idx] = 0.1 * np.nansum(SHI_temp)
+            if np.max(column_shi_elements) > 0:
+                shi[az_idx, rg_idx] = 0.1 * np.sum(column_shi_elements * dz_dataset[rg_idx])
 
     # calc maximum estimated severe hail (mm)
     if (
         mesh_method == "witt1998"
     ):  # 75th percentil fit from witt et al. 1998 (fitted to 147 reports)
-        MESH = 2.54 * SHI ** 0.5
+        mesh = 2.54 * shi ** 0.5
         mesh_description = "Maximum Estimated Size of Hail retreival developed by Witt et al. 1998 doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 "
         mesh_comment = "75th percentile fit using 147 hail reports; only valid in the first sweep"
         
     elif (
         mesh_method == "mh2019_75"
     ):  # 75th percentile fit from Muillo and Homeyer 2019 (fitted to 5897 reports)
-        MESH = 15.096 * SHI ** 0.206
+        mesh = 15.096 * shi ** 0.206
         mesh_description = "Maximum Estimated Size of Hail retreival originally developed by Witt et al. 1998 doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 and recalibrated by Murillo and Homeyer (2021) doi:10.1175/JAMC-D-20-0271.1 "
         mesh_comment = "75th percentile fit using 5897 hail reports; only valid in the first sweep"
     elif (
         mesh_method == "mh2019_95"
     ):  # 95th percentile fit from Muillo and Homeyer 2019 (fitted to 5897 reports)
-        MESH = 22.157 * SHI ** 0.212
+        mesh = 22.157 * shi ** 0.212
         mesh_description = "Maximum Estimated Size of Hail retreival originally developed by Witt et al. 1998 doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 and recalibrated by Murillo and Homeyer (2021) doi:10.1175/JAMC-D-20-0271.1 "
         mesh_comment = "95th percentile fit using 5897 hail reports; only valid in the first sweep"
     else:
         raise ValueError(
             "unknown MESH method selects, please use witt1998, mh2019_75 or mh2019_95"
         )
-
+    
     # calc warning threshold (J/m/s) NOTE: freezing height must be in km
-    WT = 57.5 * (meltlayer / 1000) - 121
+    warning_threshold = 57.5 * (meltlayer / 1000) - 121
 
     # calc probability of severe hail (POSH) (%)
-    POSH = 29 * common.safe_log(SHI / WT) + 50
-    POSH = np.real(POSH)
-    POSH[POSH < 0] = 0
-    POSH[POSH > 100] = 100
+    posh = 29 * common.safe_log(shi / warning_threshold) + 50
+    posh = np.real(posh)
+    posh[posh < 0] = 0
+    posh[posh > 100] = 100
 
-    output_fields = dict()
-    
+    #mask outside of coverage with nan
+    posh[shi_mask] = np.nan
+    shi[shi_mask] = np.nan
+    mesh[shi_mask] = np.nan
+
     # add grids to radar object
     # unpack E into cfradial representation
-    E_cfradial = np.zeros_like(data_dict[dbz_fname])
-    for i, j in enumerate(sort_idx):
-        E_cfradial[common.get_slice(j, coords_dict)] = E[i, :, :]
-
     ke_dict = {
-        "data": E_cfradial,
+        "data": hail_ke_dataset,
         "units": "Jm-2s-1",
         "long_name": "Hail Kinetic Energy",
         "description": "Hail Kinetic Energy developed by Witt et al. 1998 doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 " + 
         hail_refl_correction_description,
     }
-    output_fields[ke_fname] = ke_dict
 
-    # SHI,MESH and POSH are only valid at the surface, to represent it in pyart radar objects, insert it into the lowest sweep
-    SHI_field = np.zeros_like(data_dict[dbz_fname])
-    SHI_field[common.get_slice(0, coords_dict)] = SHI
-    SHI_dict = {
-        "data": SHI_field,
+    # SHI,MESH and POSH are only valid at the surface as a single sweep
+    shi_dict = {
+        "data": shi,
         "units": "Jm-1s-1",
         "long_name": "Severe Hail Index",
         "description": "Severe Hail Index developed by Witt et al. (1998) doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 " + 
         hail_refl_correction_description,
         "comments": "only valid in the first sweep",
     }
-    output_fields[shi_fname] = SHI_dict
 
-    MESH_field = np.zeros_like(data_dict[dbz_fname])
-    MESH_field[common.get_slice(0, coords_dict)] = MESH
-    MESH_dict = {
-        "data": MESH_field,
+    mesh_dict = {
+        "data": mesh,
         "units": "mm",
         "long_name": "Maximum Expected Size of Hail using " + mesh_method,
         "description":mesh_description + hail_refl_correction_description,
         "comments": mesh_comment,
     }
-    output_fields[mesh_fname] = MESH_dict
     
-    POSH_field = np.zeros_like(data_dict[dbz_fname])
-    POSH_field[common.get_slice(0, coords_dict)] = POSH
-    POSH_dict = {
-        "data": POSH_field,
+    posh_dict = {
+        "data": posh,
         "units": "%",
         "long_name": "Probability of Severe Hail",
         "description": "Probability of Severe Hail developed by Witt et al. (1998) doi:10.1175/1520-0434(1998)013<0286:AEHDAF>2.0.CO;2 " +
         hail_refl_correction_description,
         "comments": "only valid in the first sweep",
     }
-    output_fields[posh_fname] = POSH_dict
     
     # return output_fields dictionary
-    return output_fields
+    return ke_dict, shi_dict, mesh_dict, posh_dict
